requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-google-genai
langchain-cohere
langchain-mistralai
langchain-community
faiss-cpu
